urldata: store dns cache timeout in an int
68 years ought to be enough for most.

+++ lib/urldata.h
@@ -1743,7 +1743,7 @@ struct UserDefined {
   struct ssl_config_data proxy_ssl;  /* user defined SSL stuff for proxy */
 #endif
   struct ssl_general_config general_ssl; /* general user defined SSL stuff */
-  long dns_cache_timeout; /* DNS cache timeout */
+  int dns_cache_timeout; /* DNS cache timeout (seconds) */
   long buffer_size;      /* size of receive buffer to use */
   unsigned int upload_buffer_size; /* size of upload buffer to use,
                                       keep it >= CURL_MAX_WRITE_SIZE */
